+++ packages/grafana-ui/src/components/ValuePicker/ValuePicker.tsx
@@ -39,7 +39,13 @@ export function ValuePicker<T>({
   const [isPicking, setIsPicking] = useState(false);
 
   const buttonEl = (
-    <Button size={size || 'sm'} icon={icon || 'plus'} onClick={() => setIsPicking(true)} variant={variant}>
+    <Button
+      size={size || 'sm'}
+      icon={icon || 'plus'}
+      onClick={() => setIsPicking(true)}
+      variant={variant}
+      aria-label={selectors.components.ValuePicker.button}
+    >
       {label}
     </Button>
   );
