+++ public/app/features/dashboard/components/PanelEditor/OptionsPaneContent.tsx
@@ -3,7 +3,8 @@ import Transition from 'react-transition-group/Transition';
 import { FieldConfigSource, GrafanaTheme, PanelPlugin, SelectableValue } from '@grafana/data';
 import { DashboardModel, PanelModel } from '../../state';
 import { CustomScrollbar, Icon, Input, Select, stylesFactory, Tab, TabContent, TabsBar, useTheme } from '@grafana/ui';
-import { DefaultFieldConfigEditor, OverrideFieldConfigEditor } from './FieldConfigEditor';
+import { OverrideFieldConfigEditor } from './OverrideFieldConfigEditor';
+import { DefaultFieldConfigEditor } from './DefaultFieldConfigEditor';
 import { css } from 'emotion';
 import { PanelOptionsTab } from './PanelOptionsTab';
 import { DashNavButton } from 'app/features/dashboard/components/DashNav/DashNavButton';
